utils/apply_frame/requirements.txt
click==8.1.7
Pillow==10.2.0
tqdm==4.66.2